Moved simplification out of mvt package
* Simplification functions have been pulled out of the mvt package
and placed in thier own package.
* Removed some dead code.
* Minor formatting

Part of #224
