--- maths/point.go
@@ -1,16 +1,0 @@
-package maths
-
-/*
-type ByXYLine []Line
-
-func (t ByXYLine) Len() int      { return len(t) }
-func (t ByXYLine) Swap(i, j int) { t[i], t[j] = t[j], t[i] }
-func (t ByXYLine) Less(i, j int) bool {
-	li, lj := t[i].LeftRightMostAsLine(), t[j].LeftRightMostAsLine()
-	ret := XYOrder(li[0], lj[0])
-	if ret == 0 {
-		ret = XYOrder(li[1], lj[1])
-	}
-	return ret == -1
-}
-*/
