+++ mvt/mvt.go
@@ -3,5 +3,5 @@ package mvt
 const (
 	Version       = 2
 	DefaultExtent = 4096
-  MimeType      = "application/vnd.mapbox-vector-tile"
+	MimeType      = "application/vnd.mapbox-vector-tile"
 )
